\docType{package}
\name{ESEM_script-package}
\alias{ESEM_script}
\alias{ESEM_script-package}
\title{The ESEM_script Package}
\description{
ESEM script writter.
}
\details{
\tabular{ll}{ Package: \tab ESEM_script\cr Type: \tab
Package\cr Version: \tab 0.1\cr Date: \tab 2014-01-21\cr
Depends: \tab R (>= 2.10.0)\cr Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr }

Writes multi group ESEM invariance scripts for MPLUS for 13
model taxonomy of invariance.
}
\author{
Philip Parker \email{philip.parker@uws.edu.au}
}
\keyword{package}

