\docType{package}
\name{ESEMpackage-package}
\alias{ESEMpackage}
\alias{ESEMpackage-package}
\title{The ESEMpackage Package}
\description{
ESEM script writer.
}
\details{
\tabular{ll}{ Package: \tab ESEMpackage\cr Type: \tab
Package\cr Version: \tab 0.1\cr Date: \tab 2014-01-22\cr
Depends: \tab R (>= 2.10.0)\cr Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr }

Writes multi group ESEM invariance scripts for MPLUS for 13
model taxonomy of invariance.
}
\author{
Philip Parker \email{philip.parker@uws.edu.au}
}
\keyword{package}

