\name{simData}
\alias{simData}
\docType{data}
\title{
Simulation data from Marsh et al. in press.
}
\description{
Simple treatment by gender four group simulation data for multigroup invariance testing.
}
\usage{data(simData)}
\format{
  A data frame with 300 observations on the following 15 variables.
  \describe{
    \item{\code{X1}}{an indicator item for latent X}
    \item{\code{X2}}{an indicator item for latent X}
    \item{\code{X3}}{an indicator item for latent X}
    \item{\code{X4}}{an indicator item for latent X}
    \item{\code{X5}}{an indicator item for latent X}
    \item{\code{X6}}{an indicator item for latent X}
    \item{\code{Y1}}{an indicator item for latent Y}
    \item{\code{Y2}}{an indicator item for latent Y}
    \item{\code{Y3}}{an indicator item for latent Y}
    \item{\code{Y4}}{an indicator item for latent Y}
    \item{\code{Y5}}{an indicator item for latent Y}
    \item{\code{Y6}}{an indicator item for latent Y}
    \item{\code{TbyG}}{numeric indicator of gender by treatment group}
    \item{\code{gender}}{numeric indicator of gender}
    \item{\code{treat}}{numeric indicator of treatment group}
  }
}
\details{
Dataset used to illustrate ESEM invariance in clinical applications in the supplementary material of Marsh et al., in press.
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/24313568}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(simData)
}
\keyword{datasets}
